pyproject.toml
[project]
name = "sqm_demo"
description = "Demo for streaming computation of quadratic manifolds"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "absl-py",
  "h5py",
  "numpy",
  "jax",
  "qmpack @ git+ssh://git@github.com/Algopaul/qmpack",
  "matplotlib"
]

dynamic = ["version"]


[project.optional-dependencies]
dev = [
  "pylint>=2.6.0",
  "yapf"
]

[tool.setuptools]
py-modules = ["sqm_demo"]
